+++ specific/init_input.cpp
@@ -52,7 +52,6 @@ static XINPUT_CAPABILITIES XInputCaps;
 static DIDEVCAPS JoyCaps;
 static JOY_AXIS_RANGE JoyRanges[JoyAxisNumber];
 
-DWORD JoystickMovement = 0;
 bool JoystickVibrationEnabled = true;
 bool JoystickLedColorEnabled = true;
 
@@ -344,6 +343,31 @@ void __cdecl WinInReadKeyboard(LPVOID lpInputData) {
 }
 
 #ifdef FEATURE_INPUT_IMPROVED
+static void PovToPos(int *xPos, int *yPos, DWORD pov) {
+	int x=0, y=0;
+	// Check if D-PAD is not centered
+	if( LOWORD(pov) != 0xFFFF ) {
+		// We need here cyclic quadrilateral with a side 32.
+		// The radius of circumscribed circle is approx 23 = 16 * SQRT(2)
+		x = +23 * phd_sin(pov * PHD_360 / 36000) / PHD_IONE;
+		y = -23 * phd_cos(pov * PHD_360 / 36000) / PHD_IONE;
+		CLAMP(x, -16, 16);
+		CLAMP(y, -16, 16);
+	}
+	if( xPos ) *xPos = x;
+	if( yPos ) *yPos = y;
+}
+
+static int SelectJoyDirection(int dp, int ls, int rs, int threshold) {
+	// D-Pad has priority
+	if( ABS(dp) > threshold ) return dp;
+	// if sticks have opposite directions, just sum them
+	if( (ls^rs) < 0 ) return ls + rs;
+	// or just select the one with max value
+	if( ABS(rs) > ABS(ls) ) return rs;
+	return ls;
+}
+
 static DWORD XInputReadJoystick(int *xPos, int *yPos) {
 	if( !xPos || !yPos || XInputIndex < 0 ) return 0;
 	*xPos = *yPos = 0;
@@ -354,15 +378,25 @@ static DWORD XInputReadJoystick(int *xPos, int *yPos) {
 	if( ERROR_SUCCESS != XInputGetStateExt(XInputIndex, &state) ) {
 		return 0;
 	}
-	if( (JoystickMovement || !XINPUT_DPAD(XInputCaps.Gamepad.wButtons)) && XInputCaps.Gamepad.sThumbLX && XInputCaps.Gamepad.sThumbLY ) {
-		*xPos += 32 * state.Gamepad.sThumbLX / PHD_ONE;
-		*yPos -= 32 * state.Gamepad.sThumbLY / PHD_ONE;
-	} else if( XINPUT_DPAD(XInputCaps.Gamepad.wButtons) ) {
-		*xPos += CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_RIGHT) ? 16 : 0;
-		*xPos -= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_LEFT) ? 16 : 0;
-		*yPos += CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_DOWN) ? 16 : 0;
-		*yPos -= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_UP) ? 16 : 0;
+
+	int dx=0, dy=0, lx=0, ly=0, rx=0, ry=0;
+	if( XINPUT_DPAD(XInputCaps.Gamepad.wButtons) ) {
+		dx += CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_RIGHT) ? 16 : 0;
+		dx -= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_LEFT) ? 16 : 0;
+		dy += CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_DOWN) ? 16 : 0;
+		dy -= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_DPAD_UP) ? 16 : 0;
 	}
+	if( XInputCaps.Gamepad.sThumbLX && XInputCaps.Gamepad.sThumbLY ) {
+		lx = 16 * state.Gamepad.sThumbLX / 0x8000;
+		ly = -16 * state.Gamepad.sThumbLY / 0x8000;
+	}
+	if( XInputCaps.Gamepad.sThumbRX && XInputCaps.Gamepad.sThumbRY ) {
+		rx = 16 * state.Gamepad.sThumbRX / 0x8000;
+		ry = -16 * state.Gamepad.sThumbRY / 0x8000;
+	}
+	*xPos = SelectJoyDirection(dx, lx, rx, 8);
+	*yPos = SelectJoyDirection(dy, ly, ry, 8);
+
 	buttonStatus |= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_Y) ? 0x001 : 0;
 	buttonStatus |= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_B) ? 0x002 : 0;
 	buttonStatus |= CHK_ANY(state.Gamepad.wButtons, XINPUT_GAMEPAD_A) ? 0x004 : 0;
@@ -388,17 +422,16 @@ static DWORD RawInputReadJoystick(int *xPos, int *yPos) {
 	if( !RawInputGetState(&state) ) {
 		return 0;
 	}
-	if( JoystickMovement ) {
-		*xPos = (int)(+32.0 * state.axisLX);
-		*yPos = (int)(-32.0 * state.axisLY);
-	} else {
-		DWORD pov = state.dPad;
-		// Check if D-PAD is not centered
-		if( LOWORD(pov) != 0xFFFF ) {
-			*xPos = +16 * phd_sin(pov * PHD_360 / 36000) / PHD_IONE;
-			*yPos = -16 * phd_cos(pov * PHD_360 / 36000) / PHD_IONE;
-		}
-	}
+
+	int dx=0, dy=0, lx=0, ly=0, rx=0, ry=0;
+	PovToPos(&dx, &dy, state.dPad);
+	lx = (int)(+16.0 * state.axisLX);
+	ly = (int)(-16.0 * state.axisLY);
+	rx = (int)(+16.0 * state.axisRX);
+	ry = (int)(-16.0 * state.axisRY);
+	*xPos = SelectJoyDirection(dx, lx, rx, 8);
+	*yPos = SelectJoyDirection(dy, ly, ry, 8);
+
 	buttonStatus |= state.btnTriangle ? 0x0001 : 0;
 	buttonStatus |= state.btnCircle ? 0x0002 : 0;
 	buttonStatus |= state.btnCross ? 0x0004 : 0;
@@ -429,17 +462,25 @@ static DWORD DInputReadJoystick(int *xPos, int *yPos) {
 	{
 		if FAILED(IDID_SysJoystick->Acquire()) return 0;
 	}
-	if( (JoystickMovement || !JoyCaps.dwPOVs) && JoyCaps.dwAxes && HAS_AXIS(JoyX) && HAS_AXIS(JoyY) ) {
-		*xPos = 32 * joyState.lX / (JoyRanges[JoyX].lMax - JoyRanges[JoyX].lMin) - 16;
-		*yPos = 32 * joyState.lY / (JoyRanges[JoyY].lMax - JoyRanges[JoyY].lMin) - 16;
-	} else if( JoyCaps.dwPOVs ) {
-		DWORD pov = joyState.rgdwPOV[0];
-		// Check if D-PAD is not centered
-		if( LOWORD(pov) != 0xFFFF ) {
-			*xPos = +16 * phd_sin(pov * PHD_360 / 36000) / PHD_IONE;
-			*yPos = -16 * phd_cos(pov * PHD_360 / 36000) / PHD_IONE;
-		}
+
+	int dx=0, dy=0, lx=0, ly=0, rx=0, ry=0;
+	if( JoyCaps.dwPOVs ) {
+		PovToPos(&dx, &dy, joyState.rgdwPOV[0]);
 	}
+	if( HAS_AXIS(JoyX) && HAS_AXIS(JoyY) ) {
+		lx = 32 * joyState.lX / (JoyRanges[JoyX].lMax - JoyRanges[JoyX].lMin) - 16;
+		ly = 32 * joyState.lY / (JoyRanges[JoyY].lMax - JoyRanges[JoyY].lMin) - 16;
+	}
+	if( HAS_AXIS(JoyZ) && HAS_AXIS(JoyRZ) ) {
+		rx = 32 * joyState.lZ  / (JoyRanges[JoyZ].lMax  - JoyRanges[JoyZ].lMin)  - 16;
+		ry = 32 * joyState.lRz / (JoyRanges[JoyRZ].lMax - JoyRanges[JoyRZ].lMin) - 16;
+	} else if( HAS_AXIS(JoyRX) && HAS_AXIS(JoyRY) ) {
+		rx = 32 * joyState.lRx / (JoyRanges[JoyRX].lMax - JoyRanges[JoyRX].lMin) - 16;
+		ry = 32 * joyState.lRy / (JoyRanges[JoyRY].lMax - JoyRanges[JoyRY].lMin) - 16;
+	}
+	*xPos = SelectJoyDirection(dx, lx, rx, 8);
+	*yPos = SelectJoyDirection(dy, ly, ry, 8);
+
 	for( DWORD i=0; i<MIN(JoyCaps.dwButtons, 32); ++i ) {
 		buttonStatus |= CHK_ANY(joyState.rgbButtons[i], 0x80) ? 1<<i : 0;
 	}