+++ specific/registry.h
@@ -63,7 +63,6 @@
 #define REG_PICKUPITEM_MODE		"PickupItemMode"
 #define REG_SAVEGAME_SLOTS		"SavegameSlots"
 #define REG_SCREENSHOT_FORMAT	"ScreenshotFormat"
-#define REG_JOYSTICK_MOVEMENT	"JoystickMovement"
 #define REG_JOYSTICK_BTN_STYLE	"JoystickButtonStyle"
 
 // BOOL value names